input correction

+++ Ginásio/PersonalTrainer.h
@@ -15,9 +15,9 @@ class PersonalTrainer : public Staff {
 	std::string specializedArea;
 
 public:
-	PersonalTrainer(int age, int wage, std::string specializedArea);
-	PersonalTrainer(std::string name, int age, int wage, std::string pwd, std::string specializedArea);
-	PersonalTrainer(int id, std::string name, int age, int wage, std::string pwd, std::string specializedArea); //For file initialization
+	PersonalTrainer(int age, double wage, std::string specializedArea);
+	PersonalTrainer(std::string name, int age, double wage, std::string pwd, std::string specializedArea);
+	PersonalTrainer(int id, std::string name, int age, double wage, std::string pwd, std::string specializedArea); //For file initialization
 
 	virtual ~PersonalTrainer();
